Game/Game/Helpers/AbilityEnumHelper.cs
using System;
using System.Collections.Generic;
using Game.Models;
using System.Linq;

namespace Game.Helpers
{
    public static class AbilityEnumHelper
    {
        /// <summary>
        /// Gets the list of locations that an Item can have.
        /// Does not include the Left and Right Finger 
        /// </summary>
        public static List<string> GetListItem
        {
            get
            {
                var myList = Enum.GetNames(typeof(AbilityEnum)).ToList();
                var myReturn = myList.Where(a =>
                                            a.ToString() != AbilityEnum.Unknown.ToString()
                                            )
                                            .OrderBy(a => a)
                                            .ToList();
                return myReturn;
            }
        }

        /// <summary>
        ///  Gets the list of locations a character can use
        ///  Removes Finger for example, and allows for left and right finger
        /// </summary>
        public static List<string> GetListCharacter
        {
            get
            {
                var myList = Enum.GetNames(typeof(AbilityEnum)).ToList();
                var myReturn = myList.Where(a =>
                                           a.ToString() != AbilityEnum.Unknown.ToString()
                                            )
                                            .OrderBy(a => a)
                                            .ToList();

                return myReturn;
            }
        }

        /// <summary>
        /// Given the String for an enum, return its value.  That allows for the enums to be numbered 2,4,6 rather than 1,2,3 
        /// </summary>
        /// <param name="value"></param>
        /// <returns></returns>
        public static AbilityEnum ConvertStringToEnum(string value)
        {
            return (AbilityEnum)Enum.Parse(typeof(AbilityEnum), value);
        }

        /// <summary>
        /// If asked for a position number, return a location.  Head as 1 etc. 
        /// This compsenstates for the enum #s not being sequential, but allows for calls to the postion for random allocation etc (roll 1-7 dice and pick a item to equipt), etc... 
        /// </summary>
        /// <param name="position"></param>
        /// <returns></returns>
        public static AbilityEnum GetLocationByPosition(int position)
        {
            switch (position)
            {
                case 1:
                    return AbilityEnum.SpeedAbility;

                case 2:
                    return AbilityEnum.StrengthAbility;

                case 3:
                    return AbilityEnum.ThiccnessAbility;

                default:
                    return AbilityEnum.StaminaAbility;
            }
        }
    }
}